+++ classes/class.ilPegasusTestingTableGUI.php
@@ -1,5 +1,4 @@
 <?php
-include_once('./Services/Table/classes/class.ilTable2GUI.php');
 include_once "class.ilPegasusTesting.php";
 
 /**