+++ plugin.php
@@ -1,7 +1,7 @@
 <?php
 $id = 'sragpegasushelper';
-$version = '5.0.0';
-$ilias_min_version = '9.0.0';
-$ilias_max_version = '9.999';
+$version = '6.0.0';
+$ilias_min_version = '10.0.0';
+$ilias_max_version = '10.999';
 $responsible = 'Jakub';
 $responsible_mail = 'jakub@deepeacademy.com';